+++ src/client/app/composables/head.ts
@@ -1,11 +1,11 @@
 import { watchEffect, Ref } from 'vue'
-import { HeadConfig, SiteData } from '../../../../types/shared'
+import { HeadConfig, SiteData } from '../../shared'
 import { Route } from '../router'
 
 export function useUpdateHead(route: Route, siteDataByRouteRef: Ref<SiteData>) {
-  const metaTags: HTMLElement[] = Array.from(document.querySelectorAll('meta'))
-
+  let managedHeadTags: HTMLElement[] = []
   let isFirstUpdate = true
+
   const updateHeadTags = (newTags: HeadConfig[]) => {
     if (import.meta.env.PROD && isFirstUpdate) {
       // in production, the initial meta tags are already pre-rendered so we
@@ -13,15 +13,44 @@ export function useUpdateHead(route: Route, siteDataByRouteRef: Ref<SiteData>) {
       isFirstUpdate = false
       return
     }
-    metaTags.forEach((el) => document.head.removeChild(el))
-    metaTags.length = 0
-    if (newTags && newTags.length) {
-      newTags.forEach((headConfig) => {
-        const el = createHeadElement(headConfig)
-        document.head.appendChild(el)
-        metaTags.push(el)
-      })
+
+    const newEls: HTMLElement[] = []
+    const commonLength = Math.min(managedHeadTags.length, newTags.length)
+    for (let i = 0; i < commonLength; i++) {
+      let el = managedHeadTags[i]
+      const [tag, attrs, innerHTML = ''] = newTags[i]
+      if (el.tagName.toLocaleLowerCase() === tag) {
+        for (const key in attrs) {
+          if (el.getAttribute(key) !== attrs[key]) {
+            el.setAttribute(key, attrs[key])
+          }
+        }
+        for (let i = 0; i < el.attributes.length; i++) {
+          const name = el.attributes[i].name
+          if (!(name in attrs)) {
+            el.removeAttribute(name)
+          }
+        }
+        if (el.innerHTML !== innerHTML) {
+          el.innerHTML = innerHTML
+        }
+      } else {
+        document.head.removeChild(el)
+        el = createHeadElement(newTags[i])
+        document.head.append(el)
+      }
+      newEls.push(el)
     }
+
+    managedHeadTags
+      .slice(commonLength)
+      .forEach((el) => document.head.removeChild(el))
+    newTags.slice(commonLength).forEach((headConfig) => {
+      const el = createHeadElement(headConfig)
+      document.head.appendChild(el)
+      newEls.push(el)
+    })
+    managedHeadTags = newEls
   }
 
   watchEffect(() => {
@@ -30,25 +59,17 @@ export function useUpdateHead(route: Route, siteDataByRouteRef: Ref<SiteData>) {
     const pageTitle = pageData && pageData.title
     const pageDescription = pageData && pageData.description
     const frontmatterHead = pageData && pageData.frontmatter.head
+
+    // update title and description
     document.title = (pageTitle ? pageTitle + ` | ` : ``) + siteData.title
+    document
+      .querySelector(`meta[name=description]`)!
+      .setAttribute('content', pageDescription || siteData.description)
+
     updateHeadTags([
-      ['meta', { charset: 'utf-8' }],
-      [
-        'meta',
-        {
-          name: 'viewport',
-          content: 'width=device-width,initial-scale=1'
-        }
-      ],
-      [
-        'meta',
-        {
-          name: 'description',
-          content: pageDescription || siteData.description
-        }
-      ],
-      ...siteData.head,
-      ...((frontmatterHead && filterOutHeadDescription(frontmatterHead)) || [])
+      // site head can only change during dev
+      ...(import.meta.env.DEV ? siteData.head : []),
+      ...(frontmatterHead ? filterOutHeadDescription(frontmatterHead) : [])
     ])
   })
 }