+++ src/client/index.ts
@@ -3,19 +3,24 @@
 
 // generic types
 export type { Router, Route } from './app/router'
-
+export type { VitePressData } from './app/data'
 // theme types
 export type { Theme, EnhanceAppContext } from './app/theme'
+// shared types
+export type {
+  PageData,
+  SiteData,
+  HeadConfig,
+  Header,
+  LocaleConfig
+} from '../../types/shared'
 
 // composables
+export { useData } from './app/data'
 export { useRouter, useRoute } from './app/router'
-export { useSiteData } from './app/composables/siteData'
-export { useSiteDataByRoute } from './app/composables/siteDataByRoute'
-export { usePageData } from './app/composables/pageData'
-export { useFrontmatter } from './app/composables/frontmatter'
 
 // utilities
-export { inBrowser, joinPath } from './app/utils'
+export { inBrowser, withBase } from './app/utils'
 
 // components
 export { Content } from './app/components/Content'