+++ src/client/theme-default/components/AlgoliaSearchBox.vue
@@ -1,17 +1,10 @@
-<template>
-  <div class="algolia-search-box" id="docsearch" />
-</template>
-
 <script setup lang="ts">
 import '@docsearch/css'
-import { useRoute, useRouter } from 'vitepress'
-import { defineProps, getCurrentInstance, onMounted, watch } from 'vue'
 import docsearch from '@docsearch/js'
+import { useRoute, useRouter, useData } from 'vitepress'
+import { getCurrentInstance, onMounted, watch } from 'vue'
 import type { DefaultTheme } from '../config'
 import type { DocSearchHit } from '@docsearch/react/dist/esm/types'
-import { useSiteDataByRoute } from 'vitepress'
-
-const siteData = useSiteDataByRoute()
 
 const props = defineProps<{
   options: DefaultTheme.AlgoliaSearchOptions
@@ -57,10 +50,12 @@ function update(options: any) {
   }
 }
 
+const { lang } = useData()
+
 // if the user has multiple locales, the search results should be filtered
 // based on the language
 const facetFilters: string[] = props.multilang
-  ? ['language:' + siteData.value.lang]
+  ? ['language:' + lang.value]
   : []
 
 if (props.options.searchParameters?.facetFilters) {
@@ -68,7 +63,7 @@ if (props.options.searchParameters?.facetFilters) {
 }
 
 watch(
-  () => siteData.value.lang,
+  lang,
   (newLang, oldLang) => {
     const index = facetFilters.findIndex(
       (filter) => filter === 'language:' + oldLang
@@ -161,6 +156,10 @@ function initialize(userOptions: any) {
 }
 </script>
 
+<template>
+  <div class="algolia-search-box" id="docsearch" />
+</template>
+
 <style>
 .algolia-search-box {
   padding-top: 1px;
@@ -169,13 +168,12 @@ function initialize(userOptions: any) {
 @media (min-width: 720px) {
   .algolia-search-box {
     padding-left: 8px;
-    min-width: 176.3px; /* avoid layout shift */
   }
 }
 
 @media (min-width: 751px) {
   .algolia-search-box {
-    padding-left: 8px;
+    min-width: 176.3px; /* avoid layout shift */
   }
 
   .algolia-search-box .DocSearch-Button-Placeholder {