+++ src/client/theme-default/components/BuySellAds.vue
@@ -1,11 +1,5 @@
-<template>
-  <div class="buy-sell-ads">
-    <div class="bsa-cpc" />
-  </div>
-</template>
-
 <script setup lang="ts">
-import { defineProps, onMounted } from 'vue'
+import { onMounted } from 'vue'
 
 // global _bsa
 const ID = 'bsa-cpc-script'
@@ -51,6 +45,10 @@ onMounted(() => {
 
 function load() {
   if (typeof _bsa !== 'undefined' && _bsa) {
+    const parent = document.querySelector('.bsa-cpc')!
+    // cleanup any existing ad to avoid them stacking
+    parent.innerHTML = ''
+
     _bsa.init('default', code, `placement:${placement}`, {
       target: '.bsa-cpc',
       align: 'horizontal',
@@ -59,6 +57,12 @@ function load() {
   }
 }
 </script>
+
+<template>
+  <div class="buy-sell-ads">
+    <div class="bsa-cpc" />
+  </div>
+</template>
 
 <style scoped>
 .buy-sell-ads {