+++ src/client/theme-default/components/PageFooter.vue
@@ -1,3 +1,8 @@
+<script setup lang="ts">
+import EditLink from './EditLink.vue'
+import LastUpdated from './LastUpdated.vue'
+</script>
+
 <template>
   <footer class="page-footer">
     <div class="edit">
@@ -8,11 +13,6 @@
     </div>
   </footer>
 </template>
-
-<script setup lang="ts">
-import EditLink from './EditLink.vue'
-import LastUpdated from './LastUpdated.vue'
-</script>
 
 <style scoped>
 .page-footer {