+++ src/node/markdown/plugins/containers.ts
@@ -1,10 +1,10 @@
 import MarkdownIt from 'markdown-it'
 import Token from 'markdown-it/lib/token'
-
-const container = require('markdown-it-container')
+import container from 'markdown-it-container'
 
 export const containerPlugin = (md: MarkdownIt) => {
   md.use(...createContainer('tip', 'TIP'))
+    .use(...createContainer('info', 'INFO'))
     .use(...createContainer('warning', 'WARNING'))
     .use(...createContainer('danger', 'WARNING'))
     // explicitly escape Vue syntax