+++ src/node/markdown/plugins/link.ts
@@ -5,6 +5,7 @@
 import MarkdownIt from 'markdown-it'
 import { MarkdownParsedData } from '../markdown'
 import { URL } from 'url'
+import { EXTERNAL_URL_RE } from '../../shared'
 
 const indexRE = /(^|.*\/)index.md(#?.*)$/i
 
@@ -18,7 +19,7 @@ export const linkPlugin = (
     if (hrefIndex >= 0) {
       const hrefAttr = token.attrs![hrefIndex]
       const url = hrefAttr[1]
-      const isExternal = /^https?:/.test(url)
+      const isExternal = EXTERNAL_URL_RE.test(url)
       if (isExternal) {
         Object.entries(externalAttrs).forEach(([key, val]) => {
           token.attrSet(key, val)
@@ -31,6 +32,13 @@ export const linkPlugin = (
       ) {
         normalizeHref(hrefAttr)
       }
+
+      // encode vite-specific replace strings in case they appear in URLs
+      // this also excludes them from build-time replacements (which injects
+      // <wbr/> and will break URLs)
+      hrefAttr[1] = hrefAttr[1]
+        .replace(/\bimport\.meta/g, 'import%2Emeta')
+        .replace(/\bprocess\.env/g, 'process%2Eenv')
     }
     return self.renderToken(tokens, idx, options)
   }