+++ src/node/markdownToVue.ts
@@ -4,29 +4,48 @@ import matter from 'gray-matter'
 import LRUCache from 'lru-cache'
 import { createMarkdownRenderer, MarkdownOptions } from './markdown/markdown'
 import { deeplyParseHeader } from './utils/parseHeader'
-import { PageData, HeadConfig } from '../../types/shared'
+import { PageData, HeadConfig } from './shared'
 import { slash } from './utils/slash'
 import chalk from 'chalk'
+import _debug from 'debug'
 
-const debug = require('debug')('vitepress:md')
+const debug = _debug('vitepress:md')
 const cache = new LRUCache<string, MarkdownCompileResult>({ max: 1024 })
+const includesRE = /<!--\s*@include:\s*(.*?)\s*-->/g
 
-interface MarkdownCompileResult {
+export interface MarkdownCompileResult {
   vueSrc: string
   pageData: PageData
   deadLinks: string[]
+  includes: string[]
 }
 
 export function createMarkdownToVueRenderFn(
-  root: string,
+  srcDir: string,
   options: MarkdownOptions = {},
-  pages: string[]
+  pages: string[],
+  userDefines: Record<string, any> | undefined,
+  isBuild = false
 ) {
-  const md = createMarkdownRenderer(root, options)
+  const md = createMarkdownRenderer(srcDir, options)
   pages = pages.map((p) => slash(p.replace(/\.md$/, '')))
 
-  return (src: string, file: string): MarkdownCompileResult => {
-    const relativePath = slash(path.relative(root, file))
+  const userDefineRegex = userDefines
+    ? new RegExp(
+        `\\b(${Object.keys(userDefines)
+          .map((key) => key.replace(/[-[\]/{}()*+?.\\^$|]/g, '\\$&'))
+          .join('|')})`,
+        'g'
+      )
+    : null
+
+  return (
+    src: string,
+    file: string,
+    publicDir: string
+  ): MarkdownCompileResult => {
+    const relativePath = slash(path.relative(srcDir, file))
+    const dir = path.dirname(file)
 
     const cached = cache.get(src)
     if (cached) {
@@ -36,13 +55,33 @@ export function createMarkdownToVueRenderFn(
 
     const start = Date.now()
 
+    // resolve includes
+    let includes: string[] = []
+    src = src.replace(includesRE, (_, m1) => {
+      const includePath = path.join(dir, m1)
+      console.log(includePath)
+      const content = fs.readFileSync(includePath, 'utf-8')
+      includes.push(slash(includePath))
+      return content
+    })
+
     const { content, data: frontmatter } = matter(src)
     let { html, data } = md.render(content)
 
-    // avoid env variables being replaced by vite
-    html = html
-      .replace(/import\.meta/g, 'import.<wbr/>meta')
-      .replace(/process\.env/g, 'process.<wbr/>env')
+    if (isBuild) {
+      // avoid env variables being replaced by vite
+      html = html
+        .replace(/\bimport\.meta/g, 'import.<wbr/>meta')
+        .replace(/\bprocess\.env/g, 'process.<wbr/>env')
+
+      // also avoid replacing vite user defines
+      if (userDefineRegex) {
+        html = html.replace(
+          userDefineRegex,
+          (_) => `${_[0]}<wbr/>${_.slice(1)}`
+        )
+      }
+    }
 
     // validate data.links
     const deadLinks = []
@@ -54,9 +93,12 @@ export function createMarkdownToVueRenderFn(
         const resolved = slash(
           url.startsWith('/')
             ? url.slice(1)
-            : path.relative(root, path.resolve(dir, url))
+            : path.relative(srcDir, path.resolve(dir, url))
         )
-        if (!pages.includes(resolved)) {
+        if (
+          !pages.includes(resolved) &&
+          !fs.existsSync(path.resolve(dir, publicDir, `${resolved}.html`))
+        ) {
           console.warn(
             chalk.yellow(
               `\n(!) Found dead link ${chalk.cyan(
@@ -88,7 +130,8 @@ export function createMarkdownToVueRenderFn(
     const result = {
       vueSrc,
       pageData,
-      deadLinks
+      deadLinks,
+      includes
     }
     cache.set(src, result)
     return result
@@ -97,6 +140,7 @@ export function createMarkdownToVueRenderFn(
 
 const scriptRE = /<\/script>/
 const scriptSetupRE = /<\s*script[^>]*\bsetup\b[^>]*/
+const scriptClientRe = /<\s*script[^>]*\bclient\b[^>]*/
 const defaultExportRE = /((?:^|\n|;)\s*)export(\s*)default/
 const namedDefaultExportRE = /((?:^|\n|;)\s*)export(.+)as(\s*)default/
 
@@ -106,7 +150,11 @@ function genPageDataCode(tags: string[], data: PageData) {
   )}`
 
   const existingScriptIndex = tags.findIndex((tag) => {
-    return scriptRE.test(tag) && !scriptSetupRE.test(tag)
+    return (
+      scriptRE.test(tag) &&
+      !scriptSetupRE.test(tag) &&
+      !scriptClientRe.test(tag)
+    )
   })
 
   if (existingScriptIndex > -1) {